# climate-data-api/src/swagger.yaml
openapi: 3.0.0
info:
  title: Express API for QLD Climate Data
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: QLD Climate data server
paths:
...
  /api/user/new:
    post:
      summary: Insert a new user
      description: Creates a new user with the specified email, password, and role.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                role:
                  type: string
                  enum: [Teacher, User, Sensor]
              required:
                - email
                - password
                - role
      responses:
        "201":
          description: User created successfully
        "400":
          description: Bad request
        "401":
          description: Unauthorized - Authentication required
        "403":
          description: Forbidden - Access denied
  /api/user/access:
    put:
      summary: Update access level for users
      description: Updates the role of users created within a specified date range.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
                newRole:
                  type: string
              required:
                - startDate
                - endDate
                - newRole
      responses:
        "200":
          description: Access level updated successfully
        "400":
          description: Bad request
        "401":
          description: Unauthorized - Authentication required
        "403":
          description: Forbidden - Access denied
        "500":
          description: Server error
  /api/user/{id}:
    delete:
      summary: Delete a user by ID
      description: Deletes a single user based on the user ID.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The user ID
      responses:
        "204":
          description: User deleted successfully
        "401":
          description: Unauthorized - Authentication required
        "403":
          description: Forbidden - Access denied
        "404":
          description: User not found
        "500":
          description: Server error
  /api/user/:
    delete:
      summary: Delete multiple users
      description: Deletes multiple users who are of the "User" role and last logged in between two specified dates.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: startDate
          required: true
          schema:
            type: string
            format: date
          description: Start date of the login range
        - in: query
          name: endDate
          required: true
          schema:
            type: string
            format: date
          description: End date of the login range
      responses:
        "204":
          description: Users deleted successfully
        "401":
          description: Unauthorized - Authentication required
        "400":
          description: Bad Request - Invalid request format or parameters
        "404":
          description: Not Found - The specified resource(s) could not be found
        "403":
          description: Forbidden - Access denied
        "500":
          description: Server error

  # * WEATHER ROUTES
...